+++ packages/terra-navigation-side-menu/CHANGELOG.md
@@ -5,6 +5,8 @@
 * Added
   * Added custom status display for Drill-in content.
   * Added isLoading prop for Drill-in item.
+  * Added headerLevel prop for the title of the menu.
+  * Added missing A11y attributes.
 
 ## 2.58.1 - (May 20, 2024)