+++ packages/terra-navigation-side-menu/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * Added
+  * Added custom status display for Drill-in content.
+  * Added isLoading prop for Drill-in item.
   * Added headerLevel prop for the title of the menu.
   * Added missing A11y attributes.